+++ enigma/users/routes.py
@@ -1,4 +1,9 @@
-from flask import Blueprint
+from flask import Blueprint, render_template, url_for, flash, redirect, request
+from enigma import db, argon2
+from enigma.forms import RegistrationForm, LoginForm, UpdateAccountForm, RequestResetForm, ResetPasswordForm
+from enigma.models import User, Post
+from utils import save_picture, send_reset_email
+from flask_login import login_user, current_user, logout_user, login_required
 
 users = Blueprint("users", __name__)
 
@@ -61,7 +66,6 @@ def user_posts(username):
     user = User.query.filter_by(username=username).first_or_404()
     posts = Post.query.filter_by(author=user).order_by(Post.date_posted.desc()).paginate(page=page, per_page=5)
     return render_template('user_posts.html', posts=posts, user=user)
-
 
 @users.route("/reset_password", methods=["GET", "POST"])
 def reset_request():
