+++ enigma/users/utils.py
@@ -1,3 +1,10 @@
+import os
+import secrets
+from PIL import Image
+from flask import url_for
+from enigma import app, mail
+from flask_mail import Message
+
 def save_picture(form_picture):
     random_hex = secrets.token_hex(8)
     _, f_ext = os.path.splitext(form_picture.filename)
